fix: Use fetchGnoZenao stuff as fetchGno stuff, add feed id to gno networks
