+++ networks.json
@@ -4700,6 +4700,7 @@
     "vaultContractAddress": "",
     "nameServiceContractAddress": "gno.land/r/demo/users",
     "nameServiceDefaultImage": "ipfs://bafkreigqm3i4urywxohvpkbz5sgxosiogwn433ufmzz7eqldjxoi6xmwma",
+    "socialFeedsFeedId": "g1tas4cej0myrnyu237zgcu8wvvpqetch3k2fjf7:main",
     "socialFeedsPkgPath": "gno.land/r/zenao/social_feed",
     "socialFeedsDAOPkgPath": "gno.land/r/teritori/social_feeds_dao",
     "daoInterfacesPkgPath": "gno.land/p/teritori/dao_interfaces",
@@ -4743,6 +4744,7 @@
     "vaultContractAddress": "",
     "nameServiceContractAddress": "gno.land/r/demo/users",
     "nameServiceDefaultImage": "ipfs://bafkreigqm3i4urywxohvpkbz5sgxosiogwn433ufmzz7eqldjxoi6xmwma",
+    "socialFeedsFeedId": "g1kzqq22apmtgzd49zlxwyma0x7cu3llr968mdzn:main",
     "socialFeedsPkgPath": "gno.land/r/zenao/social_feed",
     "socialFeedsDAOPkgPath": "gno.land/r/teritori/social_feeds_dao",
     "daoInterfacesPkgPath": "gno.land/p/teritori/dao_interfaces",
